Make service more modular
In order to have an application that scales and is easy to follow
this commit splits the responsibilites of the service into
Repositories, clients and services
