+++ moj-node/server/config.js
@@ -15,6 +15,6 @@ module.exports = {
   production,
   sessionSecret: get('SESSION_SECRET', 'app-insecure-default-session', { requireInProduction: true }),
   api: {
-    hubContent: `${get('HUB_API_ENDPOINT')}/content`,
+    hubContent: `${get('HUB_API_ENDPOINT')}content`,
   },
 };
