+++ moj-node/server/index.js
@@ -5,13 +5,20 @@ const appInfoService = require('./services/appInfo');
 const createDemoDataService = require('./services/demoDataService');
 const createMenuService = require('./services/menuService');
 const createHubFeaturedContentService = require('./services/hubFeaturedContent');
+const HubClient = require('./clients/hub');
+const featuredContentRepository = require('./repositories/hubFeaturedContent');
 
 const buildInfo = config.dev ? null : require('../build-info.json'); // eslint-disable-line import/no-unresolved
 
 // pass in dependencies of service
 const demoDataService = createDemoDataService();
 const menuService = createMenuService({});
-const hubFeaturedContentService = createHubFeaturedContentService();
+
+const hubFeaturedContentService = createHubFeaturedContentService(
+  featuredContentRepository(
+    new HubClient(),
+  ),
+);
 
 const app = createApp({
   logger,
